+++ src/controllers/WebhooksController.php
@@ -10,8 +10,8 @@ namespace craft\stripe\controllers;
 use Craft;
 use craft\helpers\App;
 use craft\helpers\UrlHelper;
-use craft\stripe\models\Settings;
 use craft\stripe\Plugin;
+use craft\stripe\records\Webhook as WebhookRecord;
 use craft\web\Controller;
 use Stripe\Exception\SignatureVerificationException;
 use Stripe\Exception\UnexpectedValueException;
@@ -39,10 +39,15 @@ class WebhooksController extends Controller
         // Disable CSRF only for incoming webhooks (to agree with `allowAnonymous`, above):
         if ($action->id === 'handle') {
             $this->enableCsrfValidation = false;
+        } else {
+            // for all other requests, enforce only admins having access to the webhooks page
+            // and allow that access even if allowAdminChanges is turned off
+            $this->requireAdmin(false);
         }
 
         return parent::beforeAction($action);
     }
+
     /**
      * Handle incoming Stripe webhook event
      *
@@ -52,13 +57,11 @@ class WebhooksController extends Controller
     {
         $apiService = Plugin::getInstance()->getApi();
         $webhookService = Plugin::getInstance()->getWebhooks();
-        //$client = $apiService->getClient();
         $webhookSigningSecret = $apiService->getWebhookSigningSecret();
 
         // verify
         $payload = @file_get_contents('php://input');
         $signatureHeader = $_SERVER['HTTP_STRIPE_SIGNATURE'];
-        //$event = null;
 
         try {
             $event = Webhook::constructEvent(
@@ -76,10 +79,14 @@ class WebhooksController extends Controller
             return $this->asRaw('Err');
         }
 
+        $this->response->setStatusCode(200);
+        // as per https://docs.stripe.com/webhooks#acknowledge-events-immediately - send response asap
+        $this->response->sendAndClose();
+
         // Handle the event
         $webhookService->processEvent($event);
 
-        $this->response->setStatusCode(200);
+        // leaving this in even after sendAndClose() so that the response type matches
         return $this->asRaw('OK');
     }
 
@@ -97,9 +104,12 @@ class WebhooksController extends Controller
             throw new ServerErrorHttpException('No Stripe API key found. Make sure you have added one in the plugin’s settings screen.');
         }
 
+        $webhookRecord = $plugin->getWebhooks()->getWebhookRecord();
+
         $webhookInfo = [];
         $hasWebhook = true;
-        $webhookId = App::parseEnv($pluginSettings->webhookId);
+        $webhookId = App::parseEnv($webhookRecord->webhookId);
+        $webhookSigningSecret = App::parseEnv($webhookRecord->webhookSigningSecret);
 
         if (!empty($webhookId)) {
             try {
@@ -114,9 +124,44 @@ class WebhooksController extends Controller
         }
 
         return $this->renderTemplate('stripe/webhooks/_index', [
+            'recordId' => $webhookRecord->id,
             'webhookInfo' => $webhookInfo,
             'hasWebhook' => $hasWebhook,
+            'webhookSigningSecret' => $webhookSigningSecret,
+            'webhookId' => $webhookId,
         ]);
+    }
+
+    public function actionSaveSettings(): YiiResponse
+    {
+        $this->requirePostRequest();
+
+        $plugin = Plugin::getInstance();
+
+        // still check the API key and don't allow further actions if we don't have one
+        $pluginSettings = $plugin->getSettings();
+        if (!$pluginSettings->secretKey) {
+            throw new ServerErrorHttpException('No Stripe API key found. Make sure you have added one in the plugin’s settings screen.');
+        }
+
+        $webhookId = $this->request->getBodyParam('webhookId', null);
+        $webhookSigningSecret = $this->request->getRequiredBodyParam('webhookSigningSecret');
+
+        $webhookRecord = $plugin->getWebhooks()->getWebhookRecord();
+        $webhookRecord->webhookSigningSecret = $webhookSigningSecret;
+        $webhookRecord->webhookId = $webhookId;
+
+        if (!$webhookRecord->save()) {
+            $this->setFailFlash(Craft::t('stripe', 'Couldn’t save webhook settings.'));
+            return $this->redirectToPostedUrl();
+        }
+
+        $this->setSuccessFlash(Craft::t(
+            'stripe',
+            'Webhook settings saved.')
+        );
+
+        return $this->redirectToPostedUrl();
     }
 
     /**
@@ -158,6 +203,7 @@ class WebhooksController extends Controller
                     'customer.subscription.pending_update_expired',
                     'customer.subscription.deleted',
                     'customer.created',
+                    'customer.updated',
                     'customer.deleted',
                     'payment_method.attached',
                     'payment_method.automatically_updated',
@@ -184,8 +230,8 @@ class WebhooksController extends Controller
             return $this->redirectToPostedUrl();
         }
 
-        // get the webhook signing secret
-        $this->saveWebhookData($plugin, $pluginSettings, $response);
+        // save the webhook data (signing secret and id)
+        $this->saveWebhookData($plugin, $response);
 
         return $this->redirectToPostedUrl();
     }
@@ -198,32 +244,37 @@ class WebhooksController extends Controller
     public function actionDelete(): YiiResponse
     {
         $this->requireAcceptsJson();
+
         $id = Craft::$app->getRequest()->getBodyParam('id');
 
         $stripe = Plugin::getInstance()->getApi()->getClient();
+        $webhookRecord = WebhookRecord::findOne(['id' => $id]);
 
         try {
-            $stripe->webhookEndpoints->delete($id);
+            $stripe->webhookEndpoints->delete(App::parseEnv($webhookRecord->webhookId));
         } catch (\Exception $e) {
             Craft::error('Webhook could not be deleted: ' . $e->getMessage());
             return $this->asFailure(Craft::t('stripe', 'Webhook could not be deleted'));
         }
 
+        // delete the record from the table too
+        $webhookRecord->delete();
+
         return $this->asSuccess(Craft::t('stripe', 'Webhook deleted'));
     }
 
     /**
-     * Saves the webhook id and signing secret in the .env file and/or database and update plugin's settings.
+     * Saves the webhook id and signing secret in the .env file if possible, and in the database.
      * Sets the relevant flash message.
      *
      * @param Plugin $plugin
-     * @param Settings $settings
      * @param WebhookEndpoint $response
      * @return void
      */
-    private function saveWebhookData(Plugin $plugin, Settings $settings, WebhookEndpoint $response): void
+    private function saveWebhookData(Plugin $plugin, WebhookEndpoint $response): void
     {
         $configService = Craft::$app->getConfig();
+        $record = $plugin->getWebhooks()->getWebhookRecord();
 
         $success = true;
         try {
@@ -232,7 +283,7 @@ class WebhooksController extends Controller
             $success = false;
             Craft::error('Couldn\'t save the Stripe Webhook Signing Secret in the .env file. ' . $e->getMessage());
         }
-        $success ? $settings->webhookSigningSecret = '$STRIPE_WH_KEY' : $response->secret;
+        $record->webhookSigningSecret = $success ? '$STRIPE_WH_KEY' : $response->secret;
 
         $success = true;
         try {
@@ -241,12 +292,12 @@ class WebhooksController extends Controller
             $success = false;
             Craft::error('Couldn\'t save the Stripe Webhook ID in the .env file. ' . $e->getMessage());
         }
-        $success ? $settings->webhookId = '$STRIPE_WH_ID' : $response->id;
+        $record->webhookId = $success ? '$STRIPE_WH_ID' : $response->id;
 
-        if (!Craft::$app->getPlugins()->savePluginSettings($plugin, $settings->toArray())) {
+        if (!$record->save()) {
             Craft::$app->getSession()->setNotice(Craft::t(
                 'stripe',
-                'Webhook registered successfully, but we had trouble saving the Webhook Signing Secret. Please go to your Stripe Dashboard, get the webhook signing secret and add it to your plugin’s settings.')
+                'Webhook registered successfully, but we had trouble saving those settings. Please go to your Stripe Dashboard, get the webhook signing secret and id and add them to your plugin (Control Panel > Stripe > Webhooks).')
             );
         } else {
             $this->setSuccessFlash(Craft::t(
@@ -258,6 +309,7 @@ class WebhooksController extends Controller
 
     /**
      * Returns Webhook Endpoint info from Stripe.
+     * It's needed so that we know what the webhook can do.
      *
      * @param Plugin $plugin
      * @param string $webhookId