+++ src/services/Api.php
@@ -17,6 +17,7 @@ use craft\stripe\models\Customer;
 use craft\stripe\models\Invoice;
 use craft\stripe\models\PaymentMethod;
 use craft\stripe\Plugin;
+use Generator;
 use Stripe\Customer as StripeCustomer;
 use Stripe\Invoice as StripeInvoice;
 use Stripe\PaymentMethod as StripePaymentMethod;
@@ -128,7 +129,7 @@ class Api extends Component
 
         foreach ($customers as $customer) {
             // get user for customer's email address
-            $user = Craft::$app->getUsers()->getUserByUsernameOrEmail($customer->email);
+            $user = $customer->email ? Craft::$app->getUsers()->getUserByUsernameOrEmail($customer->email) : null;
 
             // only get payment methods if the user exists
             if ($user) {
@@ -163,13 +164,14 @@ class Api extends Component
     /**
      * Retrieve all customers.
      *
+     * @param array $params
      * @return array
      */
-    public function fetchAllCustomers(): array
+    public function fetchAllCustomers(array $params = []): array
     {
-        return $this->fetchAll('customers', [
+        return $this->fetchAll('customers', array_merge($params, [
             'expand' => $this->prepExpandForFetchAll(Customer::$expandParams),
-        ]);
+        ]));
     }
 
     /**
@@ -229,6 +231,33 @@ class Api extends Component
     }
 
     /**
+     * @param string $type
+     * @param array $params
+     * @return Generator
+     */
+    public function fetchAllIterator(string $type, array $params = []): Generator
+    {
+        $params['limit'] = 100;
+
+        $batch = $this->getClient()->$type->all($params);
+        $buffer = [];
+
+        foreach ($batch->autoPagingIterator() as $item) {
+            $buffer[] = $item;
+
+            if (count($buffer) === 100) {
+                yield $buffer;
+                $buffer = [];
+            }
+        }
+
+        // Yield any remaining items
+        if (!empty($buffer)) {
+            yield $buffer;
+        }
+    }
+
+    /**
      * Retrieves single API resource by ID.
      *
      * @param string $id Stripe ID of the object to fetch
@@ -278,13 +307,13 @@ class Api extends Component
     }
 
     /**
-     * Get parsed webhook signing secret secret
+     * Get parsed webhook signing secret
      * @return string|null
      */
     public function getWebhookSigningSecret(): ?string
     {
-        $settings = Plugin::getInstance()->getSettings();
-        return App::parseEnv($settings->webhookSigningSecret);
+        $webhookRecord = Plugin::getInstance()->getWebhooks()->getWebhookRecord();
+        return App::parseEnv($webhookRecord->webhookSigningSecret);
     }
 
     /**
@@ -295,7 +324,7 @@ class Api extends Component
      * @param array $params
      * @return array
      */
-    private function prepExpandForFetchAll(array $params): array
+    public function prepExpandForFetchAll(array $params): array
     {
         array_walk($params, fn(&$item) => $item = 'data.' . $item);